\name{df.alp-methods}
\docType{methods}
\alias{df.alp}
\title{ Extracts Deletion Diagnostics For Association Parameters}
\description{
    Extracts the deletion diagnostics for the association parameters
    which we call "dfalphas".  Works on both objects of 
    class cluster.diag.orth and obsvnl.diag.orth.
}
\section{Methods}{
    \describe{
    \item{object = "cluster.diag.orth"}{ Extracts cluster level dfalphas }
    \item{object = "obsvnl.diag.orth"}{ Extracts observational level dfalphas }
    }
}
\keyword{methods}
\seealso{\code{\link{df.bet}}}
