\name{diag.obs}
\alias{diag.obs}

\title{ Computes Observational Level Diagnostics }
\description{
    Computes observational level deletion diagnostics using as input
    an object of \bold{class} "orth" which is created by running the "orth" \bold{function}.
}

\usage{
diag.obs(object)
}


\arguments{
  \item{object}{ An object of class "orth" }
}
\details{

}
\value{
  The function returns an object of class "obsvnl.diag.orth".
}

\references{
    Preisser, J. S., Perin, J., Deletion diagnostics for marginal mean and correlation \cr
    model parameters in estimating equations, \emph{Statistics and Computing}, 17:381-393, 2007.
}

\seealso{ \code{\link{obsvnl.diag.orth-class}}, \code{\link{diag.cls}} }

\keyword{classes}