\name{formula-methods}
\docType{methods}
\alias{formula}
\title{ Extracting The Model Formula From An Orth Object}
\description{
    One of the slots of an orth object is CALL.  This CALL contains the \cr
    model formula for fitting the regression model.  This routine \cr
    extracts the formula from the CALL.  The user should not use this \cr
    routine.  The \bold{update} routine needs this routine.
}
\section{Methods}{
\describe{
\item{object = "orth"}{ Extracts cluster level dfalphas }
}}
\keyword{methods}

