\name{orth-package}
\alias{orth-package}

\docType{package}

\title{
Regression Models For Correlated Binary Data Using Orthogonalized Residuals 
or Alternating Logistic Regressions 
}

\description{
The ORTH package enables users to fit multivariate logistic regression models using the method
of orthogonalized residuals (Zink, 2003) or alternating logistic regression (Carey et al, 1993)
which is a special case of the method of orthogonalized residuals.  
}

\details{
\tabular{ll}{
Package: \tab orth\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2008-05-13\cr
License: \tab (>= 2)\cr
}


}
\author{
Written by Kunthel By, Bahjat F. Qaqish, and John S. Preisser

Maintainer: Kunthel By \email{kby@bios.unc.edu} }

\section{URL}{ \url{http://www.unc.edu/~kby/R_Splus/rs.html } }

\references{
R. C. Zink. \emph{Correlated Binary Regression Using Orthogonalized Residuals}.\cr
PhD thesis, University of North Carolina, Chapel Hill, 2003.\cr\cr

V. Carey, S. L. Zeger, P. Diggle, Modelling multivariate binary data with alternating\cr
logistic regression. \emph{Biometrika}, 80:517--526, 1993. 

}
\keyword{package}
