\name{six.city}
\alias{six.city}
\docType{data}

\title{ 6 Cities Study}
\description{
  Data from the 6 cities study.  
}

\usage{data(six.city)}
\format{
  A data frame with 128 observations on the following 5 variables.
  \describe{
    \item{\code{id}}{cluster id}
    \item{\code{y}}{response}
    \item{\code{smoke}}{smoking status}
    \item{\code{age}}{age when measurements were made}
    \item{\code{count}}{cluster weights}
  }
}
\keyword{datasets}
