\name{update-methods}
\docType{methods}
\alias{update}
\title{ Updating An Orth Object When The User Provides A Different Model Formula }
\description{
    This routine is used to update the result of a regression when the user provides\cr
    a different model formula for the marginal mean model.  If the user wants to provide \cr
    a different model formula for the association model, then \code{update} cannot be used. \cr
    The user must refit.
}
\section{Methods}{
\describe{
    \item{object = "orth"}{ Extracts cluster level dfalphas }
}}
\seealso{

	\code{\link{orth} }
}
\keyword{methods}

